Restructure api tests
