+++ web/cherrymusic/apps/api/v1/tests/views/__init__.py
@@ -1,0 +1,2 @@
+
+UNAUTHENTICATED_RESPONSE = {'detail': 'Authentication credentials were not provided.'}
